Remove analytics
At this point, the analytics add very little value to me, other than suggesting that most of my audience are indeed robots. I may decide in the future to adopt Google Analytics (a less intrusive form of tracking...and only as part of a traditional "side project" to understand how to track people), but in terms of actually understanding whether a project is working fine...it only goes to show that it doesn't.