Remove date from blog navigation screens
Users do not need to know when a blog post has been posted until
after they start reading the blog post.

Removing dates can also limit the potential damage of not having
super-recent blog posts on the front page of oyur blog, since
a "stale" blog may be considered unappealing. Removing dates
eliminates the possibility of staleness.

This is a fairly minor change though, and it is unlikely that
the target audience care about the "recency" of blog posts. This
may be because they are running their own blogs and sympathize
with the demands of content generation, or are employers who are
more interested in the writing style and insights of the blogger.

There's also a minor UI improvement from removing extraenous info,
but this is then counterbalanced by the minor UI degregation from
violating standard norms of blogs (most blogs do display dates).

Be prepared to revert this change if you find more information
indicating that "dated content" is essential on technical websites.